% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neon_nitrate}
\alias{neon_nitrate}
\title{NEON Nitrate Survey Data}
\format{
A data frame with one row per 15-minute sample:
\describe{
  \item{start_date_time}{Start time of sample (POSIXct)}
  \item{end_date_time}{End time of sample (POSIXct)}
  \item{surf_water_nitrate_mean}{Mean water nitrate concentration (mg/L)}
  \item{surf_water_nitrate_minimum}{Minimum nitrate concentration (mg/L)}
  \item{surf_water_nitrate_maximum}{Maximum nitrate concentration (mg/L)}
  \item{surf_water_nitrate_variance}{Variance of nitrate values}
  \item{surf_water_nitrate_num_pts}{Number of measurements in interval}
  \item{surf_water_nitrate_exp_uncert}{Experimental uncertainty}
  \item{surf_water_nitrate_std_er_mean}{Standard error of mean}
  \item{final_qf}{Quality flag}
}
}
\source{
NEON project: COMO site nitrate data, https://data.neonscience.org/
}
\usage{
neon_nitrate
}
\description{
Surface water nitrate data collected by NEON at COM sites.
}
\details{
Data are derived from high-frequency COM site sensors, cleaned to remove missing values in mean nitrate and timestamps.
}
\examples{
data(neon_nitrate)
summary(neon_nitrate$surf_water_nitrate_mean)
head(neon_nitrate)
}
\seealso{
[nitrate_sample_distribution], [strata_summary], [pop_est]
}
\keyword{datasets}
