% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nitrate_sample_distribution}
\alias{nitrate_sample_distribution}
\title{NEON Nitrate Sample Distribution}
\format{
A data frame:
\describe{
\item{surf_water_nitrate_mean}{Mean nitrate value (mg/L)}
\item{n}{Number of samples with that mean value}
}
}
\usage{
nitrate_sample_distribution
}
\description{
Frequency of observed surface water nitrate mean concentrations.
}
\examples{
data(nitrate_sample_distribution)
plot(nitrate_sample_distribution)
}
\keyword{datasets}
