% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_est}
\alias{pop_est}
\title{NEON Nitrate Observation Coverage}
\format{
A data frame:
\describe{
\item{total_obs}{Total number of cleaned observations}
\item{start_range}{Earliest sample time}
\item{end_range}{Latest sample time}
}
}
\usage{
pop_est
}
\description{
Dataset summarizing total number of samples and temporal coverage.
}
\examples{
data(pop_est)
pop_est$total_obs
}
\keyword{datasets}
